# WorkerThreadStart volatile read+cmpxchg loop