#Currently ThreadPoolMgr::WorkerThreadStart and ThreadPoolMgr::MaybeAddWorkingWorker use `FastInterlockCompareExchangeLong` to do read and couple with a `FastInterlockCompareExchangeLong` loop to do set.

Current Code:
```cpp
counts = WorkerCounter.GetCleanCounts(); 
// return FastInterlockCompareExchangeLong(&counts.AsLongLong,0,0);
while (true)
{
    newCounts = counts;
    // adjust newCounts to wanted values

    Counts oldCounts = WorkerCounter.CompareExchangeCounts(newCounts, counts);

    if (oldCounts == counts)
        break;

    counts = oldCounts;
}
```
It looks like the `lock cmpxchg` "write read" causes a pipeline stall. As its then followed by a `lock cmpxchg` loop to set the value it should be able to use a volatile read to get the initial value as that will then be validated by the set.

Changed code:
```cpp
counts = WorkerCounter.DangerousGetDirtyCounts(); 
// return VolatileLoad(&counts.AsLongLong);
while (true)
{
    newCounts = counts;
    // adjust newCounts to wanted values

    Counts oldCounts = WorkerCounter.CompareExchangeCounts(newCounts, counts);

    if (oldCounts == counts)
        break;

    counts = oldCounts;
}
```

Also changed GetCleanCounts to use the VolatileLoad on x64.

Before WorkerThreadStart is 2nd most expensive function and MaybeAddWorkingWorker is 10th for multicore contested QUWI
![WorkerThreadStart-1](https://aoa.blob.core.windows.net/aspnet/WorkerThreadStart-1.png)

After (3 commits) WorkerThreadStart drops to 17th and the rest of the native functions are also pushed out of the top hot spots:
![WorkerThreadStart-5](https://aoa.blob.core.windows.net/aspnet/WorkerThreadStart-5.png)

Resolves #6132
Resolves #6476

Questions:
* Is `_WIN64` the correct ifdef to used for x64? (Or `_TARGET_64BIT_` or `AMD64` etc)

/cc  @kouvel @jkotas @stephentoub