@benaadams 2016-07-29
With last commit (3) it pushes the top 9 function hotspots back into the managed stack

![WorkerThreadStart-4](https://aoa.blob.core.windows.net/aspnet/WorkerThreadStart-4.png)