@benaadams 2016-07-29
Still getting some false sharing