@benaadams 2016-07-29
![WorkerThreadStart-5](https://aoa.blob.core.windows.net/aspnet/WorkerThreadStart-5.png)