@benaadams 2016-07-29
Last commit needs a triple check that the logic is right.

It changes 3x `Volatile<T>` to normal vars +`VolatileStore` & `VolatileLoad`.

More significantly, it changes what happens when they are read and set a group. It uses 2x regular `=` and a single `VolatileX` to create a memory barrier, rather than using one for each item. 

The two spots are:

[Store fence comes first](https://github.com/benaadams/coreclr/blob/7387b825f90b15204a24d06fc40b72f961d1824e/src/vm/win32threadpool.cpp#L1256-L1259) which I believe matches [`VolatileStore in volatile.h`](https://github.com/dotnet/coreclr/blob/master/src/inc/volatile.h#L164-L174)

and:

[Load fence comes last](https://github.com/benaadams/coreclr/blob/7387b825f90b15204a24d06fc40b72f961d1824e/src/vm/win32threadpool.h#L1141-L1143) which I believe matches [`VolatileLoad in volatile.h`](https://github.com/dotnet/coreclr/blob/master/src/inc/volatile.h#L118-L129)